Updated 'notifications' project structure to existing project structure
