--- application/src/main/java/com/github/pieter_groenendijk/model/LendingAssociatedNotificationTask.java
@@ -1,37 +1,0 @@
-package com.github.pieter_groenendijk.model;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "LendingAssociatedNotificationTask")
-public class LendingAssociatedNotificationTask {
-    @EmbeddedId
-    private LendingAssociatedNotificationTaskId id;
-
-    @ManyToOne
-    @MapsId("lendingId")
-    @JoinColumn(
-        name = "lendingId",
-        nullable = false
-    )
-    private Lending lending;
-
-    @ManyToOne
-    @MapsId("notificationTaskId")
-    @JoinColumn(
-        name = "notificationTaskId",
-        nullable = false
-    )
-    private NotificationTask notificationTask;
-
-    public LendingAssociatedNotificationTask() {} // Empty constructor for ORM
-
-    public LendingAssociatedNotificationTask(Lending lending, NotificationTask notificationTask) {
-        this.lending = lending;
-        this.notificationTask = notificationTask;
-        this.id = new LendingAssociatedNotificationTaskId(
-            lending.id,
-            notificationTask.id
-        );
-    }
-}
