+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/notifiers/AppNotifier.java
@@ -1,0 +1,10 @@
+package com.github.pieter_groenendijk.service.notification.notifiers;
+
+import com.github.pieter_groenendijk.model.notification.NotificationTask;
+
+public class AppNotifier extends ChannelNotifier {
+    @Override
+    protected void attempt(NotificationTask task) throws Exception {
+
+    }
+}
