+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/notifiers/EmailNotifier.java
@@ -1,0 +1,9 @@
+package com.github.pieter_groenendijk.service.notification.notifiers;
+
+import com.github.pieter_groenendijk.model.notification.NotificationTask;
+
+public class EmailNotifier extends ChannelNotifier {
+    @Override
+    protected void attempt(NotificationTask task) throws Exception {
+    }
+}
