+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/notifiers/Notifier.java
@@ -1,0 +1,7 @@
+package com.github.pieter_groenendijk.service.notification.notifiers;
+
+import com.github.pieter_groenendijk.model.notification.NotificationTask;
+
+public interface Notifier {
+    void send(NotificationTask task);
+}
