+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/send_strategies/NotificationSendStrategyFactory.java
@@ -1,0 +1,46 @@
+package com.github.pieter_groenendijk.service.notification.send_strategies;
+
+import com.github.pieter_groenendijk.service.notification.notifiers.AppNotifier;
+import com.github.pieter_groenendijk.service.notification.notifiers.EmailNotifier;
+import com.github.pieter_groenendijk.service.notification.notifiers.SMSNotifier;
+
+public class NotificationSendStrategyFactory {
+    private final EmailNotifier emailNotifier;
+    private final SMSNotifier smsNotifier;
+    private final AppNotifier appNotifier;
+
+    public NotificationSendStrategyFactory() {
+        this(
+            new EmailNotifier(),
+            new SMSNotifier(),
+            new AppNotifier()
+        );
+    }
+
+    public NotificationSendStrategyFactory(EmailNotifier emailNotifier, SMSNotifier smsNotifier, AppNotifier appNotifier) {
+        this.emailNotifier = emailNotifier;
+        this.smsNotifier = smsNotifier;
+        this.appNotifier = appNotifier;
+    }
+
+    public AlertNotificationSendStrategy createAlertStrategy() {
+        return new AlertNotificationSendStrategy(
+            emailNotifier,
+            smsNotifier,
+            appNotifier
+        );
+    }
+
+    public WarningNotificationSendStrategy createWarningStrategy() {
+        return new WarningNotificationSendStrategy(
+            emailNotifier,
+            appNotifier
+        );
+    }
+
+    public ReminderNotificationSendStrategy createReminderStrategy() {
+        return new ReminderNotificationSendStrategy(
+            emailNotifier
+        );
+    }
+}
