+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/send_strategies/ReminderNotificationSendStrategy.java
@@ -1,0 +1,13 @@
+package com.github.pieter_groenendijk.service.notification.send_strategies;
+
+
+import com.github.pieter_groenendijk.service.notification.notifiers.EmailNotifier;
+import com.github.pieter_groenendijk.service.notification.notifiers.Notifier;
+
+public class ReminderNotificationSendStrategy extends NotificationSendStrategy {
+    public ReminderNotificationSendStrategy(EmailNotifier emailNotifier) {
+        super(new Notifier[]{
+            emailNotifier
+        });
+    }
+}
