+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/send_strategies/registry/SendStrategyType.java
@@ -1,0 +1,7 @@
+package com.github.pieter_groenendijk.service.notification.send_strategies.registry;
+
+public enum SendStrategyType {
+    ALERT,
+    WARNING,
+    REMINDER
+}
