+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/task/INotificationTaskStorage.java
@@ -1,0 +1,7 @@
+package com.github.pieter_groenendijk.service.notification.task;
+
+import com.github.pieter_groenendijk.model.notification.NotificationTask;
+
+public interface INotificationTaskStorage {
+    void store(NotificationTask task);
+}
