+++ application/src/main/java/com/github/pieter_groenendijk/service/notification/task/NotificationTaskStatus.java
@@ -1,0 +1,7 @@
+package com.github.pieter_groenendijk.service.notification.task;
+
+public enum NotificationTaskStatus {
+    PENDING,
+    EXECUTED,
+    FAILED
+}
