--- application/src/main/java/com/github/pieter_groenendijk/services/notifications/notifiers/AppNotifier.java
@@ -1,10 +1,0 @@
-package com.github.pieter_groenendijk.services.notifications.notifiers;
-
-import com.github.pieter_groenendijk.model.NotificationTask;
-
-public class AppNotifier extends ChannelNotifier {
-    @Override
-    protected void attempt(NotificationTask task) throws Exception {
-
-    }
-}
