--- application/src/main/java/com/github/pieter_groenendijk/services/notifications/notifiers/EmailNotifier.java
@@ -1,9 +1,0 @@
-package com.github.pieter_groenendijk.services.notifications.notifiers;
-
-import com.github.pieter_groenendijk.model.NotificationTask;
-
-public class EmailNotifier extends ChannelNotifier {
-    @Override
-    protected void attempt(NotificationTask task) throws Exception {
-    }
-}
