--- application/src/main/java/com/github/pieter_groenendijk/services/notifications/notifiers/Notifier.java
@@ -1,7 +1,0 @@
-package com.github.pieter_groenendijk.services.notifications.notifiers;
-
-import com.github.pieter_groenendijk.model.NotificationTask;
-
-public interface Notifier {
-    void send(NotificationTask task);
-}
