--- application/src/main/java/com/github/pieter_groenendijk/services/notifications/send_strategies/registry/NotificationSendStrategyRegistry.java
@@ -1,25 +1,0 @@
-package com.github.pieter_groenendijk.services.notifications.send_strategies.registry;
-
-import com.github.pieter_groenendijk.services.notifications.send_strategies.*;
-
-public class NotificationSendStrategyRegistry {
-    private final AlertNotificationSendStrategy alertNotificationSendStrategy;
-    private final WarningNotificationSendStrategy warningNotificationSendStrategy;
-    private final ReminderNotificationSendStrategy reminderNotificationSendStrategy;
-
-    public NotificationSendStrategyRegistry(
-        NotificationSendStrategyFactory factory
-    ) {
-        this.alertNotificationSendStrategy = factory.createAlertStrategy();
-        this.warningNotificationSendStrategy = factory.createWarningStrategy();
-        this.reminderNotificationSendStrategy = factory.createReminderStrategy();
-    }
-
-    public NotificationSendStrategy fromStrategyType(SendStrategyType type) {
-        return switch (type) {
-            case ALERT -> alertNotificationSendStrategy;
-            case WARNING -> warningNotificationSendStrategy;
-            case REMINDER -> reminderNotificationSendStrategy;
-        };
-    }
-}
