--- application/src/main/java/com/github/pieter_groenendijk/services/notifications/task/INotificationTaskStorage.java
@@ -1,7 +1,0 @@
-package com.github.pieter_groenendijk.services.notifications.task;
-
-import com.github.pieter_groenendijk.model.NotificationTask;
-
-public interface INotificationTaskStorage {
-    void store(NotificationTask task);
-}
