--- application/src/main/java/com/github/pieter_groenendijk/services/notifications/task/NotificationTaskStatus.java
@@ -1,7 +1,0 @@
-package com.github.pieter_groenendijk.services.notifications.task;
-
-public enum NotificationTaskStatus {
-    PENDING,
-    EXECUTED,
-    FAILED
-}
